src/test/java/org/smartdeveloperhub/vocabulary/util/license/Licenses.java
/**
 * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
 *   This file is part of the Smart Developer Hub Project:
 *     http://www.smartdeveloperhub.org/
 *
 *   Center for Open Middleware
 *     http://www.centeropenmiddleware.com/
 * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
 *   Copyright (C) 2015-2016 Center for Open Middleware.
 * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
 *   Licensed under the Apache License, Version 2.0 (the "License");
 *   you may not use this file except in compliance with the License.
 *   You may obtain a copy of the License at
 *
 *             http://www.apache.org/licenses/LICENSE-2.0
 *
 *   Unless required by applicable law or agreed to in writing, software
 *   distributed under the License is distributed on an "AS IS" BASIS,
 *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *   See the License for the specific language governing permissions and
 *   limitations under the License.
 * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
 *   Artifact    : org.smartdeveloperhub.vocabulary:sdh-vocabulary:0.3.0-SNAPSHOT
 *   Bundle      : sdh-vocabulary-0.3.0-SNAPSHOT.jar
 * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
 */
package org.smartdeveloperhub.vocabulary.util.license;

import java.io.IOException;
import java.net.URI;

import com.google.common.base.Optional;

public final class Licenses {

	private Licenses() {
	}

	public static License resolve(final URI identifier) throws IOException {
		try {
			final Optional<String> licenseId = LicensiusClient.findLicense(identifier.toString());
			String licenseName="unknown";
			if(licenseId.isPresent()) {
				final Optional<LicenseInfo> licenseInfo = LicensiusClient.getLicenseInfo(licenseId.orNull());
				if(licenseInfo.isPresent()) {
					final LicenseInfo info = licenseInfo.orNull();
					licenseName=info.getTitle();
					if(info.getVersion()!=null) {
						licenseName+=" "+info.getVersion();
					}
				}
			}
			return License.create(identifier, licenseName);
		} catch (final IOException e) {
			throw new IOException("Could not resolve license '"+identifier+"'",e);
		}
	}

}
